KAN-46 | Add Partners hub layout & flow - moved constants to mock-data folder
