+++ components/ambassador/ambassador-form/set/index.tsx
@@ -11,8 +11,8 @@ import { Checkbox } from "@/components/ui/Checkbox"
 import { Form, FormControl, FormField, FormItem, FormLabel } from "@/components/ui/Form"
 import { Input } from "@/components/ui/Input"
 import { Label } from "@/components/ui/Label"
-import { professionalInfo } from "@/constants/professionalInfo"
-import { socialLinks } from "@/constants/socialLinks"
+import { professionalInfo } from "@/mock-data/professionalInfo"
+import { socialLinks } from "@/mock-data/socialLinks"
 import { User } from "@/models/user"
 import documentIcon from "@/public/Document.svg"
 import { useCooperationStore } from "@/stores/cooperation"
